Initial version of the auto backup code
git-svn-id: http://192.168.1.10/repos/toto@2830 b53939d4-f08c-4181-96d4-b201aa9c61c4

+++ Toto/trunk/src/com/azquo/admin/BackupService.java
@@ -9,10 +9,12 @@ Report server logic for creating and restoring backups
 
 import com.azquo.admin.database.Database;
 import com.azquo.admin.database.DatabaseDAO;
+import com.azquo.admin.database.DatabaseServer;
 import com.azquo.admin.database.DatabaseServerDAO;
 import com.azquo.admin.onlinereport.OnlineReport;
 import com.azquo.admin.onlinereport.OnlineReportDAO;
 import com.azquo.dataimport.ImportService;
+import com.azquo.memorydb.DatabaseAccessToken;
 import com.azquo.memorydb.NameForBackup;
 import com.azquo.memorydb.ProvenanceForBackup;
 import com.azquo.memorydb.ValueForBackup;
@@ -41,94 +43,8 @@ public class BackupService {
         // ok, new code to dump a database and all reports. The former being the more difficult bit.
         File temp = File.createTempFile(loggedInUser.getDatabase().getName(), ".db");
         String tempPath = temp.getPath();
+        createDBBackupFile(loggedInUser.getDatabase().getName(), loggedInUser.getDataAccessToken(), tempPath, loggedInUser.getDatabaseServer().getIp());
         temp.deleteOnExit();
-        System.out.println("attempting to create backup file " + tempPath);
-        FileOutputStream fos = new FileOutputStream(tempPath);
-        CsvWriter csvW = new CsvWriter(fos, '\t', Charset.forName("UTF-8"));
-        csvW.setUseTextQualifier(false);
-        csvW.write(loggedInUser.getDatabase().getName());
-        csvW.endRecord();
-// we start with provenance now
-        int batchNumber = 0;
-        List<ProvenanceForBackup> provenancesForBackup =
-                RMIClient.getServerInterface(loggedInUser.getDatabaseServer().getIp()).getBatchOfProvenanceForBackup(loggedInUser.getDataAccessToken(), batchNumber);
-        while (!provenancesForBackup.isEmpty()) {
-            for (ProvenanceForBackup provenanceForBackup : provenancesForBackup) {
-                csvW.write(provenanceForBackup.getId() + "");
-                csvW.write(provenanceForBackup.getJson() + "");
-                csvW.endRecord();
-            }
-            batchNumber++;
-            provenancesForBackup =
-                    RMIClient.getServerInterface(loggedInUser.getDatabaseServer().getIp()).getBatchOfProvenanceForBackup(loggedInUser.getDataAccessToken(), batchNumber);
-        }
-        csvW.write("NAMES");
-        csvW.endRecord();
-        batchNumber = 0;
-        List<NameForBackup> namesForBackup =
-                RMIClient.getServerInterface(loggedInUser.getDatabaseServer().getIp()).getBatchOfNamesForBackup(loggedInUser.getDataAccessToken(), batchNumber);
-        while (!namesForBackup.isEmpty()) {
-            for (NameForBackup nameForBackup : namesForBackup) {
-                csvW.write(nameForBackup.getId() + "");
-                csvW.write(nameForBackup.getProvenanceId() + "");
-                // attributes could have unhelpful chars
-                csvW.write(nameForBackup.getAttributes().replace("\n", "\\\\n").replace("\r", "").replace("\t", "\\\\t"));
-                //base 64 encode the bytes
-                byte[] encodedBytes = Base64.getEncoder().encode(nameForBackup.getChildren());
-                String string64 = new String(encodedBytes);
-                csvW.write(string64);
-                csvW.write(nameForBackup.getNoParents() + "");
-                csvW.write(nameForBackup.getNoValues() + "");
-                csvW.endRecord();
-            }
-            batchNumber++;
-            namesForBackup = RMIClient.getServerInterface(loggedInUser.getDatabaseServer().getIp()).getBatchOfNamesForBackup(loggedInUser.getDataAccessToken(), batchNumber);
-        }
-        csvW.write("VALUES");
-        csvW.endRecord();
-        batchNumber = 0;
-        List<ValueForBackup> valuessForBackup =
-                RMIClient.getServerInterface(loggedInUser.getDatabaseServer().getIp()).getBatchOfValuesForBackup(loggedInUser.getDataAccessToken(), batchNumber);
-        while (!valuessForBackup.isEmpty()) {
-            for (ValueForBackup valueForBackup : valuessForBackup) {
-                csvW.write(valueForBackup.getId() + "");
-                csvW.write(valueForBackup.getProvenanceId() + "");
-                // attributes could have unhelpful chars
-                csvW.write(valueForBackup.getText());
-                //base 64 encode the bytes
-                byte[] encodedBytes = Base64.getEncoder().encode(valueForBackup.getNames());
-                String string64 = new String(encodedBytes);
-                csvW.write(string64);
-                csvW.endRecord();
-            }
-            batchNumber++;
-            valuessForBackup =
-                    RMIClient.getServerInterface(loggedInUser.getDatabaseServer().getIp()).getBatchOfValuesForBackup(loggedInUser.getDataAccessToken(), batchNumber);
-        }
-
-        csvW.write("VALUEHISTORY");
-        csvW.endRecord();
-        batchNumber = 0;
-        valuessForBackup =
-                RMIClient.getServerInterface(loggedInUser.getDatabaseServer().getIp()).getBatchOfValuesHistoryForBackup(loggedInUser.getDataAccessToken(), batchNumber);
-        while (!valuessForBackup.isEmpty()) {
-            for (ValueForBackup valueForBackup : valuessForBackup) {
-                csvW.write(valueForBackup.getId() + "");
-                csvW.write(valueForBackup.getProvenanceId() + "");
-                // attributes could have unhelpful chars
-                csvW.write(valueForBackup.getText());
-                //base 64 encode the bytes
-                byte[] encodedBytes = Base64.getEncoder().encode(valueForBackup.getNames());
-                String string64 = new String(encodedBytes);
-                csvW.write(string64);
-                csvW.endRecord();
-            }
-            batchNumber++;
-            valuessForBackup =
-                    RMIClient.getServerInterface(loggedInUser.getDatabaseServer().getIp()).getBatchOfValuesHistoryForBackup(loggedInUser.getDataAccessToken(), batchNumber);
-        }
-        csvW.close();
-        fos.close();
         // so we've got the DB backup - now gather the reports and zip 'em up
         List<OnlineReport> onlineReports = OnlineReportDAO.findForDatabaseId(loggedInUser.getDatabase().getId());
         File[] filesToPack = new File[onlineReports.size() + 1];// +1 as it's the reports + the db . . .
@@ -142,7 +58,7 @@ public class BackupService {
         File tempzip = File.createTempFile(loggedInUser.getDatabase().getName(), ".zip");
         System.out.println("temp zip " + tempzip.getPath());
         ZipUtil.packEntries(filesToPack, tempzip);
-        temp.deleteOnExit();
+        tempzip.deleteOnExit();
         return tempzip;
     }
 
@@ -156,7 +72,7 @@ public class BackupService {
         // need to load the database first
         for (File f : files) {
             if (f.getName().endsWith(".db")) {
-                loadDBBackup(loggedInUser, f, database, toReturn);
+                loadDBBackup(loggedInUser, f, database, toReturn, false);
             }
         }
         // now reports
@@ -176,7 +92,7 @@ public class BackupService {
     // should be in shared?
     static int batchSize = 100_000;
 
-    public static void loadDBBackup(LoggedInUser loggedInUser, File file, String database, StringBuilder log) {
+    public static void loadDBBackup(LoggedInUser loggedInUser, File file, String database, StringBuilder log, boolean justEmpty) {
         int line = 1;
         try {
             CsvMapper csvMapper = new CsvMapper();
@@ -195,8 +111,13 @@ public class BackupService {
             // first thing to do is delete the database and all associated reports
             Database db = DatabaseDAO.findForNameAndBusinessId(database, loggedInUser.getUser().getBusinessId());
             if (db != null) {
-                AdminService.removeDatabaseById(loggedInUser, db.getId());
-                AdminService.createDatabase(db.getName(), db.getDatabaseType(), loggedInUser, DatabaseServerDAO.findById(db.getDatabaseServerId()));
+                if (justEmpty){
+                    loggedInUser.setDatabaseWithServer(DatabaseServerDAO.findById(db.getDatabaseServerId()), db);
+                    AdminService.emptyDatabase(loggedInUser, false); // don't load the setup file!
+                } else {
+                    AdminService.removeDatabaseById(loggedInUser, db.getId());
+                    AdminService.createDatabase(db.getName(), db.getDatabaseType(), loggedInUser, DatabaseServerDAO.findById(db.getDatabaseServerId()));
+                }
             } else {
                 // still not really dealing with different servers properly
                 AdminService.createDatabase(database, "", loggedInUser, loggedInUser.getDatabaseServer());
@@ -301,4 +222,94 @@ public class BackupService {
         }
     }
 
+    public static void createDBBackupFile(String databaseName, DatabaseAccessToken databaseAccessToken, String filePath, String databaseServerIP) throws Exception {
+        System.out.println("attempting to create backup file " + filePath);
+        FileOutputStream fos = new FileOutputStream(filePath);
+        CsvWriter csvW = new CsvWriter(fos, '\t', Charset.forName("UTF-8"));
+        csvW.setUseTextQualifier(false);
+        csvW.write(databaseName);
+        csvW.endRecord();
+// we start with provenance now
+        int batchNumber = 0;
+        List<ProvenanceForBackup> provenancesForBackup =
+                RMIClient.getServerInterface(databaseServerIP).getBatchOfProvenanceForBackup(databaseAccessToken, batchNumber);
+        while (!provenancesForBackup.isEmpty()) {
+            for (ProvenanceForBackup provenanceForBackup : provenancesForBackup) {
+                csvW.write(provenanceForBackup.getId() + "");
+                csvW.write(provenanceForBackup.getJson() + "");
+                csvW.endRecord();
+            }
+            batchNumber++;
+            provenancesForBackup =
+                    RMIClient.getServerInterface(databaseServerIP).getBatchOfProvenanceForBackup(databaseAccessToken, batchNumber);
+        }
+        csvW.write("NAMES");
+        csvW.endRecord();
+        batchNumber = 0;
+        List<NameForBackup> namesForBackup =
+                RMIClient.getServerInterface(databaseServerIP).getBatchOfNamesForBackup(databaseAccessToken, batchNumber);
+        while (!namesForBackup.isEmpty()) {
+            for (NameForBackup nameForBackup : namesForBackup) {
+                csvW.write(nameForBackup.getId() + "");
+                csvW.write(nameForBackup.getProvenanceId() + "");
+                // attributes could have unhelpful chars
+                csvW.write(nameForBackup.getAttributes().replace("\n", "\\\\n").replace("\r", "").replace("\t", "\\\\t"));
+                //base 64 encode the bytes
+                byte[] encodedBytes = Base64.getEncoder().encode(nameForBackup.getChildren());
+                String string64 = new String(encodedBytes);
+                csvW.write(string64);
+                csvW.write(nameForBackup.getNoParents() + "");
+                csvW.write(nameForBackup.getNoValues() + "");
+                csvW.endRecord();
+            }
+            batchNumber++;
+            namesForBackup = RMIClient.getServerInterface(databaseServerIP).getBatchOfNamesForBackup(databaseAccessToken, batchNumber);
+        }
+        csvW.write("VALUES");
+        csvW.endRecord();
+        batchNumber = 0;
+        List<ValueForBackup> valuessForBackup =
+                RMIClient.getServerInterface(databaseServerIP).getBatchOfValuesForBackup(databaseAccessToken, batchNumber);
+        while (!valuessForBackup.isEmpty()) {
+            for (ValueForBackup valueForBackup : valuessForBackup) {
+                csvW.write(valueForBackup.getId() + "");
+                csvW.write(valueForBackup.getProvenanceId() + "");
+                // attributes could have unhelpful chars
+                csvW.write(valueForBackup.getText());
+                //base 64 encode the bytes
+                byte[] encodedBytes = Base64.getEncoder().encode(valueForBackup.getNames());
+                String string64 = new String(encodedBytes);
+                csvW.write(string64);
+                csvW.endRecord();
+            }
+            batchNumber++;
+            valuessForBackup =
+                    RMIClient.getServerInterface(databaseServerIP).getBatchOfValuesForBackup(databaseAccessToken, batchNumber);
+        }
+
+        csvW.write("VALUEHISTORY");
+        csvW.endRecord();
+        batchNumber = 0;
+        valuessForBackup =
+                RMIClient.getServerInterface(databaseServerIP).getBatchOfValuesHistoryForBackup(databaseAccessToken, batchNumber);
+        while (!valuessForBackup.isEmpty()) {
+            for (ValueForBackup valueForBackup : valuessForBackup) {
+                csvW.write(valueForBackup.getId() + "");
+                csvW.write(valueForBackup.getProvenanceId() + "");
+                // attributes could have unhelpful chars
+                csvW.write(valueForBackup.getText());
+                //base 64 encode the bytes
+                byte[] encodedBytes = Base64.getEncoder().encode(valueForBackup.getNames());
+                String string64 = new String(encodedBytes);
+                csvW.write(string64);
+                csvW.endRecord();
+            }
+            batchNumber++;
+            valuessForBackup =
+                    RMIClient.getServerInterface(databaseServerIP).getBatchOfValuesHistoryForBackup(databaseAccessToken, batchNumber);
+        }
+        csvW.close();
+        fos.close();
+
+    }
 }
